feat: portfolio/onboarding layout separation
