+++ src/app/home/start/page.tsx
@@ -1,0 +1,12 @@
+import LockerOnboarding from "@/components/LockerOnboarding";
+import Main from "@/components/Main";
+
+async function OnboardingPage() {
+	return (
+		<Main>
+			<LockerOnboarding />
+		</Main>
+	);
+}
+
+export default OnboardingPage;
