+++ src/app/layout.tsx
@@ -2,6 +2,7 @@ import "@/styles/globals.css";
 
 import { Analytics as VercelAnalytics } from "@vercel/analytics/react";
 import type { Metadata } from "next";
+import { Inter } from "next/font/google";
 import { headers } from "next/headers";
 import { type ReactNode, Suspense } from "react";
 import { cookieToInitialState } from "wagmi";
@@ -12,6 +13,8 @@ import EvmProvider from "@/providers/EvmProvider";
 import { wagmiConfig } from "@/providers/wagmiConfig";
 
 import Loading from "./loading";
+
+const inter = Inter({ subsets: ["latin"] });
 
 export const metadata: Metadata = globalMetadata;
 
@@ -27,7 +30,9 @@ export default function RootLayout({
 
 	return (
 		<html lang="en" suppressHydrationWarning>
-			<body>
+			<body
+				className={`${inter.className} flex min-h-screen w-full flex-col items-center`}
+			>
 				<Suspense
 					fallback={
 						<div className="min-h-fit">
