+++ src/app/sign-up/[[...sign-up]]/page.tsx
@@ -20,7 +20,7 @@ function SignUpPage() {
 	useEffect(() => {
 		const interval = setInterval(() => {
 			if (isSignedIn) {
-				router.push(paths.HOME);
+				router.push(paths.ONBOARDING);
 			} else {
 				clearInterval(interval);
 			}
