+++ src/data/constants/paths.ts
@@ -1,6 +1,7 @@
 export const paths = {
 	LANDING: "/",
 	HOME: "/home",
+	ONBOARDING: "/home/start",
 	ACCOUNT: "/account",
 	SIGN_IN: "/sign-in",
 	SIGN_UP: "/sign-up",
